requirements.txt
certifi==2019.9.11
chardet==3.0.4
decorator==4.4.0
facebook-wda==0.4.1
idna==2.8
numpy==1.17.2
opencv-python==4.1.1.26
Pillow==6.2.0
py==1.8.0
requests==2.22.0
retry==0.9.2
six==1.12.0
urllib3==1.25.6
